Make IdTokenEvent compatible with league/event 3
